requirements.txt
rich
simple-hexdump